move function to where it's used
