+++ corehq/apps/app_manager/suite_xml/post_process/instances.py
@@ -6,9 +6,12 @@ from django.utils.translation import gettext as _
 from memoized import memoized
 
 from corehq import toggles
-from corehq.apps.app_manager.exceptions import DuplicateInstanceIdError
+from corehq.apps.app_manager.exceptions import (
+    DuplicateInstanceIdError,
+    UnknownInstanceError,
+)
 from corehq.apps.app_manager.suite_xml.contributors import PostProcessor
-from corehq.apps.app_manager.suite_xml.xml_models import Instance, require_instances
+from corehq.apps.app_manager.suite_xml.xml_models import Instance
 from corehq.util.timer import time_method
 
 
@@ -33,7 +36,7 @@ class EntryInstances(PostProcessor):
                 unknown_instance_ids
             )
         all_instances = known_instances | custom_instances
-        require_instances(entry, instances=all_instances, instance_ids=unknown_instance_ids)
+        self.require_instances(entry, instances=all_instances, instance_ids=unknown_instance_ids)
 
     def _get_all_xpaths_for_entry(self, entry):
         relevance_by_menu, menu_by_command = self._get_menu_relevance_mapping()
@@ -123,6 +126,40 @@ class EntryInstances(PostProcessor):
     def _custom_instances_by_xmlns(self):
         return {form.xmlns: form.custom_instances for form in self.app.get_forms() if form.custom_instances}
 
+    @staticmethod
+    def require_instances(entry, instances=(), instance_ids=()):
+        used = {(instance.id, instance.src) for instance in entry.instances}
+        for instance in instances:
+            if 'remote' in instance.src:
+                continue
+            if (instance.id, instance.src) not in used:
+                entry.instances.append(
+                    # it's important to make a copy,
+                    # since these can't be reused
+                    Instance(id=instance.id, src=instance.src)
+                )
+                # make sure the first instance gets inserted
+                # right after the command
+                # once you "suggest" a placement to eulxml,
+                # it'll follow your lead and place the rest of them there too
+                if len(entry.instances) == 1:
+                    instance_node = entry.node.find('instance')
+                    command_node = entry.node.find('command')
+                    entry.node.remove(instance_node)
+                    entry.node.insert(entry.node.index(command_node) + 1, instance_node)
+        covered_ids = {instance_id for instance_id, _ in used}
+        for instance_id in instance_ids:
+            if instance_id not in covered_ids:
+                raise UnknownInstanceError(
+                    "Instance reference not recognized: {} in XPath \"{}\""
+                    # to get xpath context to show in this error message
+                    # make instance_id a unicode subclass with an xpath property
+                    .format(instance_id, getattr(instance_id, 'xpath', "(XPath Unknown)")))
+
+        sorted_instances = sorted(entry.instances, key=lambda instance: instance.id)
+        if sorted_instances != entry.instances:
+            entry.instances = sorted_instances
+
 
 def get_instance_factory(scheme):
     return get_instance_factory._factory_map.get(scheme, preset_instances)
@@ -173,7 +210,9 @@ def commcare_fixture_instances(app, instance_name):
 
 
 def _commcare_reports_instances(app, instance_name, prefix):
-    from corehq.apps.app_manager.suite_xml.features.mobile_ucr import get_uuids_by_instance_id
+    from corehq.apps.app_manager.suite_xml.features.mobile_ucr import (
+        get_uuids_by_instance_id,
+    )
     if instance_name.startswith(prefix) and toggles.MOBILE_UCR.enabled(app.domain):
         instance_id = instance_name[len(prefix):]
         uuid = get_uuids_by_instance_id(app).get(instance_id, [instance_id])[0]
